Place object at the top when possible ⏫
The only interesting width monotonicity for placement is widening - because shrinking means that if it does not fit at the top, it will also not fit at the bottom. Thus, there is no need for min and max, but one can simply try to place at the top and if it doesn't fit, but does fit at the bottom then min=top and max=bot.